refactor: types

+++ src/utils/api.ts
@@ -13,10 +13,11 @@ type Car = {
     id?: number
 }
 type Winner = {
-    id?: number,
-    wins?: number,
-    time?: number
+    id: number,
+    wins: number,
+    time: number
 }
+
 
 export const generateQueryParams = (params: Param[] = []) => {
     return params ? `?${params.map((param) => `${param.key}=${param.value}`).join('&')}` : '';
@@ -27,20 +28,18 @@ export const getCars = async(queryParams: Param[] = []) => {
     const response = await fetch(`${mainUrl}${paths.cars}${generateQueryParams(queryParams)}`);
     const cars = await response.json();
     const count = Number(response.headers.get('X-Total-Count'));
-
     return { cars, count }
 }
 
 // get car by id
-export const getCar = async(id: number) => {
+export const getCar = async(id: number): Promise<Car> => {
     const response = await fetch(`${mainUrl}${paths.cars}/${id}`);
     const car = await response.json();
-
     return car;
 }
 
 // create a new car
-export const createCar = async(car: Car) => {
+export const createCar = async(car: Car): Promise<Car> => {
     const response = await fetch(`${mainUrl}${paths.cars}`, {
         method: 'POST',
         headers: {
@@ -49,12 +48,11 @@ export const createCar = async(car: Car) => {
         body: JSON.stringify(car)
     });
     const newCar = await response.json();
-
     return newCar;
 }
 
 // update car with PUT
-export const updateCar = async(id: number, car: Car) => {
+export const updateCar = async(id: number, car: Car): Promise<Car> => {
     const response = await fetch(`${mainUrl}${paths.cars}/${id}`, {
         method: 'PUT',
         headers: {
@@ -63,12 +61,11 @@ export const updateCar = async(id: number, car: Car) => {
         body: JSON.stringify(car)
     });
     const updatedCar = await response.json();
-
     return updatedCar;
 }
 
 // update car's props with PATCH
-export const patchCar = async(id: number, car: Car) => {
+export const patchCar = async(id: number, car: Car): Promise<Car> => {
     const response = await fetch(`${mainUrl}${paths.cars}/${id}`, {
         method: 'PATCH',
         headers: {
@@ -77,17 +74,15 @@ export const patchCar = async(id: number, car: Car) => {
         body: JSON.stringify(car)
     });
     const updatedCar = await response.json();
-
     return updatedCar;
 }
 
 // delete a car by id
-export const deleteCar = async(id: number) => {
+export const deleteCar = async(id: number): Promise<Car> => {
     const response = await fetch(`${mainUrl}${paths.cars}/${id}`, {
         method: 'DELETE'
     });
     const deletedCar = await response.json();
-
     return deletedCar;
 }
 
@@ -102,7 +97,6 @@ export const controlEngine = async(queryParams: Param[] = []) => {
         }
     });
     const carInfo = await response.json();
-
     return carInfo;
 }
 
@@ -113,8 +107,6 @@ export const driveEngine = async(queryParams: Param[] = []) => {
             'Content-Type': 'application/json'
         }
     });
-    // const carInfo = await response.json();
-
     return response.status;
 }
 
@@ -125,18 +117,17 @@ export const getWinners = async(queryParams: Param[] = []) => {
     const response = await fetch(`${mainUrl}${paths.winners}${generateQueryParams(queryParams)}`);
     const winners = await response.json();
     const count = Number(response.headers.get('X-Total-Count'));
-
     return { winners, count }
 }
 
-export const getWinner = async(id: number) => {
+export const getWinner = async(id: number): Promise<Winner> => {
     const response = await fetch(`${mainUrl}${paths.winners}/${id}`);
     const winner = await response.json();
     return winner;
 }
 
 // create a winner
-export const createWinner = async(winner: Winner) => {
+export const createWinner = async(winner: Winner): Promise<Winner> => {
     const response = await fetch(`${mainUrl}${paths.winners}`, {
         method: 'POST',
         headers: {
@@ -145,22 +136,24 @@ export const createWinner = async(winner: Winner) => {
         body: JSON.stringify(winner)
     });
     const newWinner = await response.json();
-
     return newWinner;
 }
 
 // delete a winner by id
 export const deleteWinner = async(id: number) => {
+    const winnerStatus = await getWinnerStatus(id);
+    if(winnerStatus === 404) {
+        return;
+    }
     const response = await fetch(`${mainUrl}${paths.winners}/${id}`, {
         method: 'DELETE'
     });
     const deletedWinner = await response.json();
-
     return deletedWinner;
 }
 
 // update winner's props with PATCH
-export const updateWinner = async(id: number, winner: Winner) => {
+export const updateWinner = async(id: number, winner: Winner): Promise<Winner> => {
     const response = await fetch(`${mainUrl}${paths.winners}/${id}`, {
         method: 'PATCH',
         headers: {
@@ -169,7 +162,6 @@ export const updateWinner = async(id: number, winner: Winner) => {
         body: JSON.stringify(winner)
     });
     const updatedWinner = await response.json();
-
     return updatedWinner;
 }
 
@@ -178,17 +170,13 @@ export const getWinnerStatus = async(id: number) => (await fetch(`${mainUrl}${pa
 export const saveWinner = async(id: number, time: number) => {
     const winnerStatus = await getWinnerStatus(id);
     if(winnerStatus === 404) {
-        await createWinner({
-            id,
-            wins: 1,
-            time
-        })
+        await createWinner({ id, wins: 1, time });
     } else {
         const winner = await getWinner(id);
         await updateWinner(id, {
             id,
             wins: winner.wins + 1,
             time: time < winner.time ? time : winner.time
-        })
+        });
     }
 }
